refactor(openomni): budget nagging is an opinion (#606)
Second move of D5. The core owns budget *accounting* — the limits are
loop invariants, and `checkBudget` is the query over them — but telling a
model how much room it has left is an opinion about how to talk to a
model. `builtin:budget-reassurance` and `builtin:budget-warning` move to
openomni and register through the public `PolicyRegistry.register`.

`checkBudget` and `describeBudgetRemaining` become exported from
`@openomni/agent`. That is the shape the split implies: the core keeps
the accounting and hands out the reading; the product decides what to say
about it.

The two snapshot blocks in `builtin-snapshots.test.ts` move with the
policies, and this time the pins that block carried travel intact —
`effectCapabilities` by deep equality, plus `policyId` and the reason
codes, which the moved behavior suite did not cover. #625's review caught
exactly that gap on the last move; the same three mutations are verified
red-first here.
