+++ packages/agent/src/core/policy/builtin/index.ts
@@ -1,6 +1,2 @@
-export {
-  createBudgetReassurancePolicy,
-  createBudgetWarningPolicy,
-} from "./budget";
 export { createCompactionPolicy } from "./compaction";
 export { createToolPermissionPolicy } from "./tool-guard";
