+++ packages/agent/src/index.ts
@@ -11,6 +11,9 @@ export type {
   Sink,
 } from "./core/types";
 export { PolicyEngine, PolicyRegistry, defaultRegistry } from "./core/policy";
+// Budget accounting stays core (the limits are loop invariants); the queries
+// are exported so a product can decide what to say about what is left (D5).
+export { checkBudget, describeBudgetRemaining } from "./core/budget";
 export type {
   CanonicalPolicyRegistration,
   PolicyContext,
@@ -26,10 +29,6 @@ export type {
 } from "./core/policy";
 export { McpClient } from "./runtime/mcp/index";
 export type { McpServerConfig } from "./runtime/mcp/index";
-export {
-  createBudgetReassurancePolicy,
-  createBudgetWarningPolicy,
-} from "./core/policy/builtin/budget";
 export { createCompactionPolicy } from "./core/policy/builtin/compaction";
 export { createToolPermissionPolicy } from "./core/policy/builtin/tool-guard";
 export { InMemoryCompactor } from "./core/execution/compaction";
